refactor: 네이밍 변경 orderConfirm -> orderCheck
