+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { BrowserRouter, Route, Routes } from 'react-router';
 import CartPage from './pages/cart/CartPage';
-import OrderConfirmPage from './pages/orderConfirm/OrderConfirmPage';
+import OrderCheckPage from './pages/orderCheck/OrderCheckPage';
 
 function App() {
   return (
     <BrowserRouter basename="/react-shopping-cart">
       <Routes>
         <Route path="/" element={<CartPage />} />
-        <Route path="/order-confirmation" element={<OrderConfirmPage />} />
+        <Route path="/order-check" element={<OrderCheckPage />} />
       </Routes>
     </BrowserRouter>
   );
